Update date of amendments
We discussed the amendments from the fall 2019 meeting during a PyHC
telecon on 06-Jan-2020, and approved the changes.  We still need to
update the list of authors of the amendment.

One topic of discussion during this telecon was on adoption of NEP 29
(which created a community-wide standard for how long to support older
versions of Python and SciPy) by PyHC packages.  Stuart Mumford
suggested that it be okay for some packages to have stricter
standards, and that packages can support older versions of NumPy and
Python if they so choose.  Michael Hirsch brought up that
usage/download statistics from a package that supports Python 3.5+
showed that very few people are using Python 3.5 at the moment (and
also that Python 3.5 gives the most trouble for the testing matrix).
I brought up that core packages like SunPy should follow the
guidelines of NEP 29 most closely, while packages that are currently
under development and have few users would be okay in supporting
Python 3.7+ or even 3.8+ if the development phase is going to last
~1-2 years.  We decided that using "should" leaves open enough leeway
for us to accept packages that have different requirements.  I forgot
to mention that I think that we should follow the same guidelines for
Astropy (and ideally packages within PyHC) as we do for NumPy.

(These notes are a convolution of what people actually said, and my
ability to remember 42 minutes into the past.)